sag-service/src/main/java/com/sagag/services/service/coupon/revalidation/CouponReValidateDateStep.java
package com.sagag.services.service.coupon.revalidation;

import com.sagag.services.service.coupon.CouponValidationData;
import com.sagag.services.service.coupon.validator.CouponDateValidator;
import com.sagag.services.service.coupon.validator.CouponValidator;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Component;

@Component
public class CouponReValidateDateStep extends AbstractCouponReValidationStep {

  @Autowired
  private CouponDateValidator dateValidator;

  @Override
  public CouponValidationData handle(CouponValidationData data) {
    return handleAsDefault(data);
  }

  @Override
  protected CouponValidator getValidator() {
    return dateValidator;
  }

  @Override
  protected String getStepName() {
    return "Coupon revalidate date step";
  }
}
